Minor tweaking.

+++ src/draw_scene.cpp
@@ -529,8 +529,8 @@ void render_scene(const level& lvl, const screen_position& pos) {
 		//first draw black over the sections of the screen which aren't to be drawn to
 		//GLshort varray1[8] = {0,0,  pixels,0,  pixels,fb->h,   0,fb->h};
 		//GLshort varray2[8] = {fb->w - pixels,0,  fb->w,0,   fb->w,fb->h,  fb->w - pixels,fb->h};
-		GLshort varray1[8] = {0,0,  pixels,0,  pixels,preferences::actual_screen_height(),   0,preferences::actual_screen_height()};
-		GLshort varray2[8] = {preferences::actual_screen_width() - pixels,0,  preferences::actual_screen_width(),0,   preferences::actual_screen_width(),preferences::actual_screen_height(),  preferences::actual_screen_width() - pixels,preferences::actual_screen_height()};
+		GLshort varray1[8] = {0,0,  GLshort(pixels),0,  GLshort(pixels),GLshort(preferences::actual_screen_height()),   0, GLshort(preferences::actual_screen_height())};
+		GLshort varray2[8] = {GLshort(preferences::actual_screen_width() - pixels),0,  GLshort(preferences::actual_screen_width()),0,  GLshort(preferences::actual_screen_width()), GLshort(preferences::actual_screen_height()),  GLshort(preferences::actual_screen_width() - pixels), GLshort(preferences::actual_screen_height())};
 		glColor4ub(0, 0, 0, 255);
 #if defined(USE_GLES2)
 		gles2::manager gles2_manager(gles2::get_simple_shader());
